+++ src/charts/fixtures/column-chart.js
@@ -1,7 +1,7 @@
 export default [
-  { y: 231, name: ['Immediate', ''] },
-  { y: 91, name: ['Jan 15, 2019—', 'Jan 28, 2019'] },
-  { y: 321, name: ['Feb 15, 2019—', 'Mar 28, 2019'] },
-  { y: 81, name: ['Apr 15, 2019—', 'May 28, 2019'] },
-  { y: 263, name: ['Jun 15, 2019—', 'Jul 28, 2019'] },
+  { y: 231, name: 'Immediate' },
+  { y: 91, name: 'Jan 15, 2019 — Jan 28, 2019' },
+  { y: 321, name: 'Feb 15, 2019 — Mar 28, 2019' },
+  { y: 81, name: 'Apr 15, 2019 — May 28, 2019' },
+  { y: 263, name: 'Jun 15, 2019 — Jul 28, 2019' },
 ];